POLIO-2125: reorganise vaccine stock api code
