+++ plugins/polio/api/vaccines/stock_management/__init__.py
@@ -1,0 +1,45 @@
+from plugins.polio.api.vaccines.stock_management.campaign import CampaignCategory, compute_category_from_campaign
+from plugins.polio.api.vaccines.stock_management.destructions import DestructionReportViewSet
+from plugins.polio.api.vaccines.stock_management.earmarks import EarmarkedStockViewSet
+from plugins.polio.api.vaccines.stock_management.incidents import IncidentReportViewSet
+from plugins.polio.api.vaccines.stock_management.outgoingstockmovement import (
+    OutgoingStockMovementPatchSerializer,
+    OutgoingStockMovementSerializer,
+    OutgoingStockMovementStrictSerializer,
+    OutgoingStockMovementViewSet,
+)
+from plugins.polio.api.vaccines.stock_management.subitems import (
+    VaccineStockSubitemBase,
+    VaccineStockSubitemEdit,
+    vaccine_stock_id_param,
+)
+from plugins.polio.api.vaccines.stock_management.vaccine_stock import (
+    EmbeddedVaccineStockManagementViewset,
+    StockManagementCustomFilter,
+    VaccineStockCreateSerializer,
+    VaccineStockListSerializer,
+    VaccineStockManagementViewSet,
+    VaccineStockSerializer,
+)
+
+
+__all__ = [
+    "CampaignCategory",
+    "DestructionReportViewSet",
+    "EmbeddedVaccineStockManagementViewset",
+    "EarmarkedStockViewSet",
+    "IncidentReportViewSet",
+    "OutgoingStockMovementPatchSerializer",
+    "OutgoingStockMovementSerializer",
+    "OutgoingStockMovementStrictSerializer",
+    "OutgoingStockMovementViewSet",
+    "StockManagementCustomFilter",
+    "VaccineStockCreateSerializer",
+    "VaccineStockListSerializer",
+    "VaccineStockManagementViewSet",
+    "VaccineStockSerializer",
+    "VaccineStockSubitemBase",
+    "VaccineStockSubitemEdit",
+    "compute_category_from_campaign",
+    "vaccine_stock_id_param",
+]
