+++ plugins/polio/api/vaccines/stock_management/destructions/__init__.py
@@ -1,0 +1,5 @@
+from plugins.polio.api.vaccines.stock_management.destructions.serializers import DestructionReportSerializer
+from plugins.polio.api.vaccines.stock_management.destructions.views import DestructionReportViewSet
+
+
+__all__ = ["DestructionReportSerializer", "DestructionReportViewSet"]
