+++ plugins/polio/api/vaccines/stock_management/incidents/__init__.py
@@ -1,0 +1,5 @@
+from plugins.polio.api.vaccines.stock_management.incidents.serializers import IncidentReportSerializer
+from plugins.polio.api.vaccines.stock_management.incidents.views import IncidentReportViewSet
+
+
+__all__ = ["IncidentReportSerializer", "IncidentReportViewSet"]
