+++ plugins/polio/api/vaccines/stock_management/outgoingstockmovement/__init__.py
@@ -1,0 +1,14 @@
+from plugins.polio.api.vaccines.stock_management.outgoingstockmovement.serializers import (
+    OutgoingStockMovementPatchSerializer,
+    OutgoingStockMovementSerializer,
+    OutgoingStockMovementStrictSerializer,
+)
+from plugins.polio.api.vaccines.stock_management.outgoingstockmovement.views import OutgoingStockMovementViewSet
+
+
+__all__ = [
+    "OutgoingStockMovementPatchSerializer",
+    "OutgoingStockMovementSerializer",
+    "OutgoingStockMovementStrictSerializer",
+    "OutgoingStockMovementViewSet",
+]
