+++ plugins/polio/api/vaccines/stock_management/vaccine_stock/__init__.py
@@ -1,0 +1,20 @@
+from plugins.polio.api.vaccines.stock_management.vaccine_stock.filters import StockManagementCustomFilter
+from plugins.polio.api.vaccines.stock_management.vaccine_stock.serializers import (
+    VaccineStockCreateSerializer,
+    VaccineStockListSerializer,
+    VaccineStockSerializer,
+)
+from plugins.polio.api.vaccines.stock_management.vaccine_stock.views import (
+    EmbeddedVaccineStockManagementViewset,
+    VaccineStockManagementViewSet,
+)
+
+
+__all__ = [
+    "EmbeddedVaccineStockManagementViewset",
+    "StockManagementCustomFilter",
+    "VaccineStockCreateSerializer",
+    "VaccineStockListSerializer",
+    "VaccineStockManagementViewSet",
+    "VaccineStockSerializer",
+]
